safari-600.7-branch/LayoutTests/js/for-in-to-text-expected.txt

This tests that JavaScriptCore ForInNodes are converted correctly to text.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS test.toString().match('for *[(]j *in *index[)]') != null is true
PASS successfullyParsed is true

TEST COMPLETE
